pyproject.toml
[project]
name = "pixlstash"
version = "0.9.3"
description = "Self-hosted picture library with AI tagging, semantic search, and a clean web UI"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Gaute Lindkvist" }
]
keywords = [
    "image library",
    "AI tagging",
    "semantic search",
    "photo management",
    "self-hosted",
    "comfyui",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "open_clip_torch",
    "fastapi",
    "uvicorn",
    "uvicorn[standard]",
    "setuptools",
    "wheel",
    "twine",
    "numpy",
    "pillow",
    "opencv-python",
    "scipy",
    "platformdirs",
    "tomli",
    "colorlog",
    "httpx",
    "python-multipart",
    "onnxruntime",
    "requests",
    "transformers<4.49",
    "torch",
    "torchvision",
    "insightface",
    "rapidfuzz",
    "tqdm",
    "einops",
    "sentence_transformers",
    "spacy",
    "pillow-heif",
    "sqlmodel",
    "alembic",
    "python-jose[cryptography]",
    "dotenv",
    "passlib",
    "bcrypt<4.0.0",
    "nvidia-ml-py",
    "piexif",
]

[project.urls]
Homepage = "https://pixelurgy.github.io/pixlstash"
Repository = "https://github.com/Pixelurgy/pixlstash"
"Bug Tracker" = "https://github.com/Pixelurgy/pixlstash/issues"
Changelog = "https://github.com/Pixelurgy/pixlstash/releases"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "psutil"
]
dev = [
    "ruff"
]

[project.scripts]
pixlstash-server = "pixlstash.app:main"

[build-system]
requires = ["setuptools>=42", "wheel", "twine"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["pixlstash*"]

[tool.setuptools.package-data]
pixlstash = [
    "alembic.ini",
    "frontend/dist/**/*",
    "migrations/*.py",
    "migrations/*.mako",
    "migrations/versions/*.py",
    "image_plugins/built-in/*.py",
    "data/anchors/*.npy",
    "data/comfyui-workflows/built-in/*.json",
]
